# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Shared Music API",
    "version": "1.0.0",
    "description": "API to share music with friends using Spotify."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user.",
        "description": "Allows a new user to register, creating a new user record in the database.",
        "tags": [
          "Registration"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The user's username."
                  },
                  "password": {
                    "type": "string",
                    "description": "The user's password."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully."
          },
          "400": {
            "description": "This user already exists."
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Authenticate a user and return a token.",
        "description": "Verifies user's credentials and returns a JWT token for authenticated sessions.",
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The user's username."
                  },
                  "password": {
                    "type": "string",
                    "description": "The user's password."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful, JWT token returned.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT token for the authenticated session."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User not found or wrong password."
          }
        }
      }
    },
    "/joinGroup": {
      "post": {
        "summary": "Join or create a group.",
        "description": "Allows the user to join a group specified by its name. If the group does not exist, it is created. If the user was already a member of another group, he is removed from it.",
        "tags": [
          "Groups"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the group to join or create."
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a message confirming that the user has successfully joined the group, or that he was already a member of it."
          },
          "400": {
            "description": "Invalid request if the group name is missing."
          },
          "500": {
            "description": "Error server if an unexpected error occurs."
          }
        }
      }
    },
    "/getGroupList": {
      "get": {
        "summary": "Obtain a list of all available groups.",
        "description": "returns a list of groups with their names and the number of members.",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of groups with their names and the number of members.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "The name of the group."
                      },
                      "numberOfUsers": {
                        "type": "integer",
                        "description": "The number of members in the group."
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error server if an unexpected error occurs."
          }
        }
      }
    },
    "/getGroupMembers": {
      "get": {
        "summary": "Obtain a list of all members of the group to which the current user belongs.",
        "description": "Returns a list of all members of the group to which the current user belongs, including information about being the leader of the group, and the currently listened Spotify track for each member if they are connected to Spotify and currently playing.",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all members of the group to which the current user belongs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "username": {
                        "type": "string",
                        "description": "The username of the member."
                      },
                      "isLeader": {
                        "type": "boolean",
                        "description": "Tells if the member is the leader of the group."
                      },
                      "currentTrack": {
                        "type": "string",
                        "description": "The song currently played by the member on Spotify, if available."
                      },
                      "device": {
                        "type": "string",
                        "description": "The device currently used by the member on Spotify, if available."
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The user does not belong to any group."
          }
        }
      }
    },
...
  },
...
}